other int -> long conversion and adena max count checks
git-svn-id: http://svn.l2jserver.com/trunk@3154 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2_GameServer/java/net/sf/l2j/gameserver/RecipeController.java
@@ -775,7 +775,7 @@ public class RecipeController
 			for (L2RecipeInstance recipe : recipes)
 			{
 				int quantity = _recipeList.isConsumable() ? (int) (recipe.getQuantity() * Config.RATE_CONSUMABLE_COST)
-				                                         : (int) recipe.getQuantity();
+				                                         : recipe.getQuantity();
 
 				if (quantity > 0)
 				{
